+++ quiki/quiki/.quiki/classes/User.php
@@ -6,7 +6,11 @@ class User {
 	protected $email;
 	protected $registered;
 
-	function __construct($name) {
+	function __construct($name)
+	{
+		// initialize database handlers
+		User::init();
+		
 		// get data for this object
 		User::$readQuery->execute(array($name));
 		if (!($data = User::$readQuery->fetch()))
@@ -18,7 +22,8 @@ class User {
 			$this->$key = $value;
 	}
 	
-	public function update($name = null, $password = null, $email = null) {
+	public function update($name = null, $password = null, $email = null)
+	{
 		// get database name
 		$dbName = $this->name;
 		
@@ -34,12 +39,14 @@ class User {
 		User::$updateQuery->execute(array($this->name, $this->password, $this->email, $dbName));
 	}
 	
-	public function delete() {
+	public function delete()
+	{
 		// delete this object
 		return (bool) User::$deleteQuery->execute(array($this->name));
 	}
 	
-	function __get($key) {
+	function __get($key)
+	{
 		// return protected values
 		if (isset($this->$key))
 			return $this->$key;
@@ -50,7 +57,11 @@ class User {
 	// static functions
 	//----------------------------------------------------------------------
 	
-	static public function create($name, $password, $email) {
+	static public function create($name, $password, $email)
+	{
+		// initialize database handlers
+		User::init();
+		
 		// validate data
 		if (!preg_match('/^[A-Za-z_][A-Za-z_0-9\-]{0,254}$/', $name))
 			throw new Exception('Please enter a valid username. A username must start with a letter, and can only contain alphanumeric characters, except for `_` and `-`.');
@@ -64,7 +75,11 @@ class User {
 		return new User($name);
 	}
 	
-	static public function register($name, $email) {
+	static public function register($name, $email)
+	{
+		// initialize database handlers
+		User::init();
+		
 		// check if a user with this username exists
 		if (count(User::getList(array('name' => $name))))
 			throw new Exception('A user with the username "' . $name . '" already exists.');
@@ -95,7 +110,11 @@ class User {
 		return $password;
 	}
 	
-	static public function getList($search = array(), $start = 0, $limit = 999999) {
+	static public function getList($search = array(), $start = 0, $limit = 999999)
+	{
+		// initialize database handlers
+		User::init();
+		
 		// create where clause
 		$where = '';
 		foreach (array_keys($search) as $key)
@@ -124,16 +143,21 @@ class User {
 	protected static $deleteQuery;
 	
 	static public function init() {
+		// initialized flag
+		static $init = false;
+		if ($init)
+			return;
+			
 		// initialize database handlers
 		User::$dbh = Quiki::getDBConnection();
 		User::$createQuery = User::$dbh->prepare('INSERT INTO users (name, password, email, registered) VALUES (?, MD5(?), ?, NOW())');
 		User::$readQuery = User::$dbh->prepare('SELECT *, UNIX_TIMESTAMP(registered) FROM `users` WHERE `name` = ?');
 		User::$updateQuery = User::$dbh->prepare('UPDATE `users` SET `name` = ?, `password` = ?, `email` = ? WHERE `name` = ?');
 		User::$deleteQuery = User::$dbh->prepare('DELETE FROM `users` WHERE `name` = ? LIMIT 1');
+	
+		// set init flag
+		return ($init = true);
 	}
 }
-
-// initialize database handlers
-User::init();
 
 ?>
